requirements.txt
customtkinter==5.2.2
lxml==5.4.0
openpyxl==3.1.5
pandas==2.2.2
PyYAML==6.0.1
pytest
pytest-mock